test: move example repo generator to `utils`
